✅  Sort out testing for classic and glp variations
